+++ web/src/auth/index.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "../hooks/tokenProvider";
+import { AuthContext } from "../hooks/authProvider";
 
 const useAuth = () => {
   return useContext(AuthContext);
